+++ csunisa/lmm.py
@@ -1,3 +1,5 @@
+import sys
+from math import factorial
 import numpy as np
 from numpy.polynomial import Polynomial
 import matplotlib.pyplot as plt
@@ -16,7 +18,7 @@ class LinearMultistepMethod:
     beta : array_like of float
         Coefficients of the f terms (β₀ to β_k).
     name : str, optional
-        Name or label for the method.
+        Label for identification.
 
     Attributes
     ----------
@@ -28,6 +30,10 @@ class LinearMultistepMethod:
         Number of steps (derived from the length of alpha).
     name : str
         Label for identification.
+    order : int
+        Consinstency order p.
+    error_constant : float
+        The first nonzero C_q constant.
     """
 
     def __init__(self, alpha, beta, name=""):
@@ -35,6 +41,8 @@ class LinearMultistepMethod:
         self.beta = np.array(beta, dtype=float)
         self.k = len(alpha) - 1
         self.name = name or "unnamed LMM"
+        self._order = None
+        self._error_constant = None
 
         # Normalise if αₖ ≠ 1
         if self.alpha[-1] != 1:
@@ -59,7 +67,38 @@ class LinearMultistepMethod:
         rho, sigma = self.characteristic_polynomials()
         return rho - hbar * sigma
 
-    def solve(self, ivp, h, ignition=[], tol=1e-6, max_iter=100):
+    def _compute_order_and_error(self):
+        q = 1
+
+        def C(q):
+            return sum(
+                j**q / factorial(q) * self.alpha[j]
+                - j**(q - 1) / factorial(q - 1) * self.beta[j]
+                for j in range(self.k + 1)
+            )
+
+        eps = np.finfo(np.float32).eps
+        constant = C(q)
+        while abs(constant) < eps:
+            q += 1
+            constant = C(q)
+
+        self._order = q - 1
+        self._error_constant = constant
+
+    @property
+    def order(self):
+        if self._order is None:
+            self._compute_order_and_error()
+        return self._order
+
+    @property
+    def error_constant(self):
+        if self._error_constant is None:
+            self._compute_order_and_error()
+        return self._error_constant
+
+    def solve(self, ivp, h, starting=[], tol=1e-6, max_iter=100):
         """
         Apply the method to an IVP.
 
@@ -68,8 +107,8 @@ class LinearMultistepMethod:
         ivp : InitialValueProblem
         h : float
             Step size.
-        ignition: ndarray (k-1, len(y0)), optional
-            Ignition values for multistep methods.
+        starting: ndarray (k-1, len(y0)), optional
+            Starting values for multistep methods.
         tol : float, default 1e-6
             Tolerance for implicit methods.
         max_iter : int, default 100
@@ -82,26 +121,29 @@ class LinearMultistepMethod:
         y : ndarray (len(t), len(y0))
             Array of solution values at each time step.
         """
-        f = ivp.f
         t0, tf = ivp.t_span
         t = np.arange(t0, tf + h, h)
         y = np.zeros((len(t), len(ivp.y0)))
         y[0] = ivp.y0
 
-        # Additional ignition values
-        if ignition.shape != (self.k-1, len(ivp.y0)):
+        # Additional starting values
+        if len(starting) == 0:
+            if self.k > 1:
+                raise ValueError(f"{self.k}-step method needs {self.k-1} "
+                                 "additional starting values (got 0)")
+        elif starting.shape != (self.k-1, len(ivp.y0)):
             raise ValueError(f"{self.k}-step method needs {self.k-1} "
-                             "additional ignition values "
-                             f"(got {ignition.shape[0]})")
+                             "additional starting values "
+                             f"(got {starting.shape[0]})")
         for i in range(1, self.k):
-            y[i] = ignition[i-1]
+            y[i] = starting[i-1]
 
         # Main loop for explicit methods
         if self.beta[-1] == 0:
             for i in range(self.k, len(t)):
                 y[i] = sum(
                     -self.alpha[j] * y[i-self.k+j]
-                    + h * self.beta[j] * f(t[i-self.k+j], y[i-self.k+j])
+                    + h * self.beta[j] * ivp.f(t[i-self.k+j], y[i-self.k+j])
                     for j in range(0, self.k)
                 )
         # Main loop for implicit methods
@@ -109,10 +151,10 @@ class LinearMultistepMethod:
             for i in range(self.k, len(t)):
                 def g(z):
                     return sum(
-                        -self.alpha[j] * y[i-self.k+j]
-                        + h * self.beta[j] * f(t[i-self.k+j], y[i-self.k+j])
+                        - self.alpha[j] * y[i-self.k+j]
+                        + h * self.beta[j] * ivp.f(t[i-self.k+j], y[i-self.k+j])
                         for j in range(0, self.k)
-                    ) + h * self.beta[self.k] * f(t[i], z)
+                    ) + h * self.beta[self.k] * ivp.f(t[i], z)
 
                 y[i], _ = fixed_point_iteration(g, y[i-1], tol, max_iter)
 
@@ -123,7 +165,6 @@ class LinearMultistepMethod:
         Plot the method's absolute stability region by applying the scanning
         technique to the square xs × ys.
         """
-
         plt.xlim(xs)
         plt.ylim(ys)
         xs = np.linspace(xs[0], xs[1], num)