+++ run_tests.py
@@ -1,11 +1,11 @@
-import tests.problems as ivps
+import unit_tests.problems as ivps
 
-from csunisa.linear_multistep_method import LinearMultistepMethod
+from csunisa.lmm import LinearMultistepMethod
 from csunisa.lmm_registry import get_method, list_methods
-from tests.tester import test_lmm
+from unit_tests.tester import test_lmm
 
 
-def unit_test_lmms(h=0.05):
+def unit_test_lmms(h=0.01):
     methods = list_methods()
 
     for method in methods:
@@ -13,7 +13,7 @@ def unit_test_lmms(h=0.05):
         lmm = LinearMultistepMethod(data["alpha"], data["beta"], method)
         print(f"🧪 Running: {lmm.name}")
         print("-" * 50)
-        for pb_function in ivps.problems.values():
+        for pb_function in ivps.lmm_problems.values():
             pb = pb_function()
             test_lmm(lmm, pb, h)
         print("\n")